Instance Registry changes for WebApi
