+++ src/main/scala/de/upb/cs/swt/delphi/instancemanagement/Instance.scala
@@ -27,7 +27,6 @@ final case class Instance (
                             host: String,
                             portNumber: Int,
                             name: String,
-                            /* Component Type */
                             componentType: InstanceEnums.ComponentType
                           )
